pyproject.toml
[tool.poetry]
name = "asam-qc-openmaterial-3d"
version = "1.0.0"
description = "This project implements ASAM OpenMATERIAL 3D quality checker."
authors = ["Clemens Linnhoff <clemens.linnhoff@persival.de>"]
license = "MPL-2.0"
readme = "README.md"
packages = [
    { include = "qc_openmaterial3d" },
]


[tool.poetry.dependencies]
python = "^3.10"
asam-qc-baselib = "^1.0.0rc1"
lxml = "^5.2.2"
jsonschema = "^4.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
black = "^24.4.2"

[tool.poetry.scripts]
qc_openmaterial = 'qc_openmaterial3d.main:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
